move config into own file and allow legacy changelog entries
