+++ CHANGELOG.md
@@ -246,7 +246,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (vesting) [#1672](https://github.com/evmos/evmos/pull/1672) Port `vesting` precompile code and refactor integration tests.
 - (vesting) [#1667](https://github.com/evmos/evmos/pull/1667) Add support for vesting precompile in the `evm` module.
 - (cli) [#1647](https://github.com/evmos/evmos/pull/1647) Update defaults on `evmosd start` flags.
-- (vesting) Refactor vesting flow
+- (vesting) Refactor vesting flow.
 
 ### Bug Fixes
