+++ scripts/changelog_checker/entry.py
@@ -5,111 +5,11 @@ relate to the changes in a specific pull request.
 The expected structure of an entry is: `- (category) [#PR](link) description`
 """
 
-import os
 import re
 from typing import Dict, List, Tuple
 
-def get_allowed_categories() -> List[str]:
-    """
-    Returns a list of allowed categories for an individual changelog entry. 
+from config import ALLOWED_CATEGORIES, ALLOWED_SPELLINGS, KNOWN_EXCEPTIONS
 
-    It is using a set of predefined categories, that are extended by the entries in
-     - the `x/...` modules
-     - the `precompiles/...` subdirectories
-     - the `precompiles/outposts/...` subdirectories
-    
-    :return: a list of allowed categories for an individual changelog entry
-    """
-
-    allowed_categories = [
-        "all",
-        "ante",
-        "api",
-        "app",
-        "build",
-        "ci",
-        "cli",
-        "crisis",
-        "db",
-        "deps",
-        "docs",
-        "docker",
-        "eip712",
-        "fees",
-        "go"
-        "make",
-        "metrics",
-        "outposts",
-        "post",
-        "precompiles",
-        "proto",
-        "release",
-        "rpc",
-        "swagger",
-        "testnet",
-        "tests",
-        "types",
-        "utils",
-        "upgrade",
-
-        # third party modules
-        "bank",
-        "distribution",
-        "gov",
-        "ics20",
-        "staking",
-
-        # outdated modules (we have to keep them since they're in the changelog)
-        "claims",
-        "consensus",
-        "recovery",
-        "incentives",
-    ]
-
-    base_path = os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
-
-    module_path = os.path.join(base_path, "x")
-    for module in os.listdir(module_path):
-        if os.path.isdir(os.path.join(module_path, module)):
-            allowed_categories.append(module)
-
-    precompile_path = os.path.join(base_path, "precompiles")
-    for precompile in os.listdir(precompile_path):
-        if os.path.isdir(os.path.join(precompile_path, precompile)):
-            allowed_categories.append(precompile + "-precompile")
-
-    outpost_path = os.path.join(base_path, "precompiles", "outposts")
-    for outpost in os.listdir(outpost_path):
-        if os.path.isdir(os.path.join(outpost_path, outpost)):
-            allowed_categories.append(outpost + "-outpost")
-
-    return allowed_categories
-
-
-# List of allowed categories for an individual changelog entry.
-ALLOWED_CATEGORIES = get_allowed_categories()
-
-# A dictionary of allowed spellings for some common patterns in changelog entries.
-ALLOWED_SPELLINGS = {
-    "ABI": re.compile("abi", re.IGNORECASE),
-    "API": re.compile("api", re.IGNORECASE),
-    "CI": re.compile("ci", re.IGNORECASE),
-    "Cosmos-SDK": re.compile(r"cosmos[\s-]*sdk", re.IGNORECASE),
-    "CLI": re.compile("cli", re.IGNORECASE),
-    "EIP-712": re.compile(r"eip[\s-]*712", re.IGNORECASE),
-    "ERC-20": re.compile(r"erc[\s-]*20", re.IGNORECASE),
-    "EVM": re.compile("evm", re.IGNORECASE),
-    "IBC": re.compile("ibc", re.IGNORECASE),
-    "ICS": re.compile("ics", re.IGNORECASE),
-    "ICS-20": re.compile(r"ics[\s-]*20", re.IGNORECASE),
-    "outpost": re.compile("outpost", re.IGNORECASE),
-    "Osmosis": re.compile("osmosis", re.IGNORECASE),
-    "PR": re.compile(r"(pr)(\s|$)", re.IGNORECASE),
-    "precompile": re.compile("precompile", re.IGNORECASE),
-    "SDK": re.compile("sdk", re.IGNORECASE),
-    "Stride": re.compile("stride", re.IGNORECASE),
-    "WERC-20": re.compile(r"werc[\s-]*20", re.IGNORECASE),
-}
 
 # Allowed entry pattern: `- (category) [#PR](link) - description`
 ENTRY_PATTERN = re.compile(
@@ -130,6 +30,7 @@ class Entry:
         self.backslash: bool = False
         self.category: str = ""
         self.description: str = ""
+        self.is_exception: bool = False
         self.link: str = ""
         self.pr_number: int = 0
         self.problems: List[str] = []
@@ -145,6 +46,11 @@ class Entry:
         problems: List[str] = []
         match = ENTRY_PATTERN.match(self.line)
         if not match:
+            if self.line in KNOWN_EXCEPTIONS:
+                self.is_exception = True
+
+                return True
+
             problems.append(f'Malformed entry: "{self.line}"')
             self.problems = problems
             return False
