+++ scripts/changelog_checker/test_entry.py
@@ -7,7 +7,6 @@ from entry import (
     check_link,
     check_spelling,
     check_whitespace,
-    get_allowed_categories,
     get_match,
     Entry,
 )
@@ -45,7 +44,6 @@ class TestEntry:
         assert ok is True
         assert entry.fixed == example
 
-    
     def test_fail_has_backslash_in_link(self):
         example = r"- (evm) [\#1851](https://github.com/evmos/evmos/pull/1851) Test."
         entry = Entry(example)
@@ -55,7 +53,6 @@ class TestEntry:
         ]
         assert ok is False
         assert entry.fixed == example.replace(r'\#', "#")
-
 
     def test_entry_wrong_pr_link_and_missing_dot(self):
         entry = Entry(
@@ -216,11 +213,3 @@ class TestGetMatch:
     def test_fail_in_link(self):
         assert get_match(re.compile("abi", re.IGNORECASE), "Fix [abcdef](https://example/aBi.com).") == ""
 
-
-class TestGetAllowedCategories:
-    def test_pass(self):
-        allowed_categories = get_allowed_categories()
-        assert "app" in allowed_categories, "expected pre-configured value to be in allowed categories"
-        assert "evm" in allowed_categories, "expected module to be in allowed categories"
-        assert "osmosis-outpost" in allowed_categories, "expected outpost to be in allowed categories"
-        assert "distribution-precompile" in allowed_categories, "expected precompile to be in allowed categories"
