+++ scripts/changelog_checker/testdata/changelog_fail.md
@@ -44,3 +44,4 @@ Some comments at head of file...
 ### API Breaking
 
 - (vesting) [#1862](https://github.com/evmos/evmos/pull/1862) Add Authorization Grants to the Vesting extension.
+- malformed entry in changelog
