+++ scripts/changelog_checker/testdata/changelog_ok.md
@@ -10,6 +10,7 @@ Some comments at head of file...
 - (p256-precompile) [#1922](https://github.com/evmos/evmos/pull/1922) Add `secp256r1` curve precompile.
 - (distribution-precompile) [#1949](https://github.com/evmos/evmos/pull/1949) Add `ClaimRewards` custom transaction.
 - (swagger) [#2218](https://github.com/evmos/evmos/pull/2218) Use correct version of proto dependencies to generate swagger.
+- (go) [#1687](https://github.com/evmos/evmos/pull/1687) Bump Evmos version to v14.
 
 ### API Breaking
 
@@ -35,3 +36,9 @@ Some comments at head of file...
 
 - (vesting) [#1862](https://github.com/evmos/evmos/pull/1862) Add Authorization Grants to the Vesting extension.
 - (app) [#555](https://github.com/evmos/evmos/pull/555) `v4.0.0` upgrade logic.
+
+## [v2.0.0](https://github.com/evmos/evmos/releases/tag/v2.0.0) - 2021-10-31
+
+### State Machine Breaking
+
+- legacy entries do not have to be fully correct
